Extract code from Elasticsearch::Index
Elasticsearch::Index is a very large class with too many
responsibilities. Delegating things to extracted classes we can start
to pull apart the unit tests, which are testing (or stubbing) too much
at the moment, making it difficult to add functionality.

`test_should_allow_custom_timeouts_on_add` had to be removed because it
is testing the internals of the class.
